# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DeclarationStatus = typeof DeclarationStatus[keyof typeof DeclarationStatus];


export const DeclarationStatus = {
  passed: 'passed',
  warning: 'warning',
  failed: 'failed',
} as const;

export interface Declaration {
  key: string;
  label: string;
  value: string;
  status: DeclarationStatus;
  confidence: number;
  requirement: string;
}

export type FindingSeverity = typeof FindingSeverity[keyof typeof FindingSeverity];


export const FindingSeverity = {
  critical: 'critical',
  major: 'major',
  minor: 'minor',
} as const;

export type FindingStatus = typeof FindingStatus[keyof typeof FindingStatus];


export const FindingStatus = {
  open: 'open',
  resolved: 'resolved',
} as const;

export interface Finding {
  id: number;
  severity: FindingSeverity;
  title: string;
  detail: string;
  rule: string;
  status: FindingStatus;
}

export interface BarcodeProduct {
  /** @nullable */
  productName: string | null;
  /** @nullable */
  manufacturer: string | null;
  /** @nullable */
  netQuantity: string | null;
  /** @nullable */
  brand: string | null;
  /** @nullable */
  category: string | null;
  source: string;
}

export interface BarcodeLookup {
  rawValue: string;
  normalizedValue: string;
  format: string;
  valid: boolean;
  found: boolean;
  product: BarcodeProduct | null;
  message: string;
}

export type ScanStatus = typeof ScanStatus[keyof typeof ScanStatus];


export const ScanStatus = {
  compliant: 'compliant',
  review: 'review',
  violation: 'violation',
} as const;

export interface Scan {
  id: number;
  productName: string;
  brand: string;
  /** @nullable */
  manufacturer: string | null;
  /** @nullable */
  netQuantity: string | null;
  category: string;
  /** @nullable */
  imageName: string | null;
  status: ScanStatus;
  riskScore: number;
  issueCount: number;
  /** @nullable */
  topIssue: string | null;
  scannedAt: string;
  inspector: string;
}

export type ScanDetail = Scan & {
  declarations: Declaration[];
  findings: Finding[];
};

export interface ScanInput {
  /** @minLength 1 */
  productName: string;
  /** @minLength 1 */
  brand: string;
  /** @nullable */
  manufacturer?: string | null;
  /** @nullable */
  netQuantity?: string | null;
  /** @minLength 1 */
  category: string;
  /** @nullable */
  imageName?: string | null;
}

export interface DashboardSummary {
  totalScans: number;
  compliantCount: number;
  reviewCount: number;
  violationCount: number;
  complianceRate: number;
  openFindings: number;
  scansToday: number;
  trend: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  scan: 'scan',
  finding: 'finding',
  report: 'report',
} as const;

export type ActivityItemStatus = typeof ActivityItemStatus[keyof typeof ActivityItemStatus];


export const ActivityItemStatus = {
  compliant: 'compliant',
  review: 'review',
  violation: 'violation',
  neutral: 'neutral',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  description: string;
  timestamp: string;
  status: ActivityItemStatus;
  /** @nullable */
  scanId: number | null;
}

export type ListScansParams = {
q?: string;
status?: ListScansStatus;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListScansStatus = typeof ListScansStatus[keyof typeof ListScansStatus];


export const ListScansStatus = {
  compliant: 'compliant',
  review: 'review',
  violation: 'violation',
  all: 'all',
} as const;