# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemStatus } from './activityItemStatus';
import type { ActivityItemType } from './activityItemType';

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  description: string;
  timestamp: Date;
  status: ActivityItemStatus;
  /** @nullable */
  scanId: number | null;
}