# lib/api-zod/src/generated/types/barcodeLookup.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BarcodeProduct } from './barcodeProduct';

export interface BarcodeLookup {
  rawValue: string;
  normalizedValue: string;
  format: string;
  valid: boolean;
  found: boolean;
  product: BarcodeProduct | null;
  message: string;
}