# lib/api-zod/src/generated/types/barcodeProduct.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BarcodeProduct {
  /** @nullable */
  productName: string | null;
  /** @nullable */
  manufacturer: string | null;
  /** @nullable */
  netQuantity: string | null;
  /** @nullable */
  brand: string | null;
  /** @nullable */
  category: string | null;
  source: string;
}