# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalScans: number;
  compliantCount: number;
  reviewCount: number;
  violationCount: number;
  complianceRate: number;
  openFindings: number;
  scansToday: number;
  trend: number;
}