# lib/api-zod/src/generated/types/declaration.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeclarationStatus } from './declarationStatus';

export interface Declaration {
  key: string;
  label: string;
  value: string;
  status: DeclarationStatus;
  confidence: number;
  requirement: string;
}