# lib/api-zod/src/generated/types/scan.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanStatus } from './scanStatus';

export interface Scan {
  id: number;
  productName: string;
  brand: string;
  /** @nullable */
  manufacturer: string | null;
  /** @nullable */
  netQuantity: string | null;
  category: string;
  /** @nullable */
  imageName: string | null;
  status: ScanStatus;
  riskScore: number;
  issueCount: number;
  /** @nullable */
  topIssue: string | null;
  scannedAt: Date;
  inspector: string;
}