# lib/api-zod/src/generated/types/scanDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Declaration } from './declaration';
import type { Finding } from './finding';
import type { Scan } from './scan';

export type ScanDetail = Scan & {
  declarations: Declaration[];
  findings: Finding[];
};