# lib/api-zod/src/generated/types/scanInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScanInput {
  /** @minLength 1 */
  productName: string;
  /** @minLength 1 */
  brand: string;
  /** @nullable */
  manufacturer?: string | null;
  /** @nullable */
  netQuantity?: string | null;
  /** @minLength 1 */
  category: string;
  /** @nullable */
  imageName?: string | null;
}